Time: 6 ms (37.97%), Space: 54.5 MB (66.94%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -1,6 +1,6 @@
 class Solution {
     public double findMaxAverage(int[] nums, int k) {
-		double sum = 0.0;
+		double sum = 0;
 		for(int i=0; i<k; i++) {
 			sum += (double) nums[i];
 		}
@@ -11,6 +11,6 @@ class Solution {
 			maxAvg = Math.max(maxAvg, sum/k);
 		}
 		
-		return maxAvg;
+		return (double) maxAvg;
     }
 }
